# Compute hash outside of critical sections