This migrates from using std::collections::HashMap to hashbrown::HashTable, which allows us to compute the hash of our strings outside of a critical section, which was suggested in #3.