@erickt 2025-07-16
Benchmarks before:

<details>

```
from_str_first/0        time:   [7.2869 ns 7.4096 ns 7.5368 ns]
Found 15 outliers among 100 measurements (15.00%)
  3 (3.00%) high mild
  12 (12.00%) high severe
from_str_first/1        time:   [13.671 ns 13.679 ns 13.688 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
from_str_first/7        time:   [6.9576 ns 7.0272 ns 7.1094 ns]
from_str_first/8        time:   [41.326 ns 41.361 ns 41.400 ns]
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
from_str_first/90       time:   [57.389 ns 57.423 ns 57.457 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
from_str_first/972      time:   [230.86 ns 230.96 ns 231.07 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low severe
  1 (1.00%) high mild

from_str_dupe/0         time:   [7.1752 ns 7.1948 ns 7.2191 ns]
Found 17 outliers among 100 measurements (17.00%)
  8 (8.00%) high mild
  9 (9.00%) high severe
from_str_dupe/1         time:   [13.680 ns 13.690 ns 13.700 ns]
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
from_str_dupe/7         time:   [6.9879 ns 7.0813 ns 7.1843 ns]
from_str_dupe/8         time:   [15.719 ns 15.728 ns 15.737 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
from_str_dupe/90        time:   [25.819 ns 25.833 ns 25.848 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
from_str_dupe/972       time:   [117.55 ns 117.60 ns 117.66 ns]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

clone/0                 time:   [3.7301 ns 3.7325 ns 3.7351 ns]
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  1 (1.00%) high severe
clone/1                 time:   [3.7459 ns 3.7497 ns 3.7537 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
clone/7                 time:   [3.7319 ns 3.7335 ns 3.7352 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
clone/8                 time:   [4.5744 ns 4.5954 ns 4.6199 ns]
clone/90                time:   [4.5409 ns 4.5515 ns 4.5626 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild
clone/972               time:   [4.5106 ns 4.5211 ns 4.5325 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild

drop_dupe/0             time:   [30.769 ns 30.786 ns 30.803 ns]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  2 (2.00%) high severe
drop_dupe/1             time:   [30.741 ns 30.760 ns 30.780 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe
drop_dupe/7             time:   [30.745 ns 30.767 ns 30.792 ns]
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
drop_dupe/8             time:   [34.197 ns 34.222 ns 34.249 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
drop_dupe/90            time:   [34.191 ns 34.215 ns 34.239 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
drop_dupe/972           time:   [34.196 ns 34.219 ns 34.243 ns]
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

drop_last/0             time:   [30.788 ns 30.809 ns 30.832 ns]
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe
drop_last/1             time:   [30.782 ns 30.806 ns 30.832 ns]
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) low mild
  9 (9.00%) high mild
  2 (2.00%) high severe
drop_last/7             time:   [30.773 ns 30.797 ns 30.822 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
drop_last/8             time:   [51.289 ns 51.336 ns 51.385 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
drop_last/90            time:   [58.083 ns 58.142 ns 58.203 ns]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
drop_last/972           time:   [139.76 ns 139.87 ns 139.98 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  4 (4.00%) high mild
  1 (1.00%) high severe

as_str/0                time:   [311.00 ps 311.43 ps 311.95 ps]
Found 16 outliers among 100 measurements (16.00%)
  8 (8.00%) high mild
  8 (8.00%) high severe
as_str/1                time:   [310.51 ps 310.73 ps 310.99 ps]
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) high mild
  7 (7.00%) high severe
as_str/7                time:   [311.78 ps 312.25 ps 312.76 ps]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
as_str/8                time:   [311.47 ps 311.88 ps 312.33 ps]
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe
as_str/90               time:   [311.86 ps 312.31 ps 312.83 ps]
Found 14 outliers among 100 measurements (14.00%)
  7 (7.00%) high mild
  7 (7.00%) high severe
as_str/972              time:   [311.62 ps 312.08 ps 312.61 ps]
Found 17 outliers among 100 measurements (17.00%)
  5 (5.00%) high mild
  12 (12.00%) high severe

ahash/0                 time:   [31.059 ns 31.083 ns 31.111 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
ahash/1                 time:   [31.136 ns 31.167 ns 31.200 ns]
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  5 (5.00%) high mild
ahash/7                 time:   [31.075 ns 31.109 ns 31.145 ns]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
ahash/8                 time:   [31.064 ns 31.088 ns 31.112 ns]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe
ahash/90                time:   [31.043 ns 31.069 ns 31.098 ns]
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe
ahash/972               time:   [31.058 ns 31.079 ns 31.101 ns]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

eq/0                    time:   [310.58 ps 310.86 ps 311.17 ps]
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
eq/1                    time:   [311.57 ps 311.86 ps 312.17 ps]
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
eq/7                    time:   [311.07 ps 311.35 ps 311.67 ps]
Found 14 outliers among 100 measurements (14.00%)
  7 (7.00%) high mild
  7 (7.00%) high severe
eq/8                    time:   [310.52 ps 310.75 ps 310.99 ps]
Found 18 outliers among 100 measurements (18.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  11 (11.00%) high severe
eq/90                   time:   [310.90 ps 311.11 ps 311.33 ps]
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
eq/972                  time:   [312.24 ps 312.64 ps 313.04 ps]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```

</details>

Benchmarks after:

<details>

```
from_str_first/0        time:   [7.2248 ns 7.2818 ns 7.3527 ns]
Found 23 outliers among 100 measurements (23.00%)
  3 (3.00%) high mild
  20 (20.00%) high severe
Benchmarking from_str_first/1: Warming up for 3.0000 sls
from_str_first/1        time:   [13.636 ns 13.652 ns 13.667 ns]
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low severe
  7 (7.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
from_str_first/7        time:   [6.9258 ns 6.9958 ns 7.0758 ns]
Found 11 outliers among 100 measurements (11.00%)
  11 (11.00%) high mild
from_str_first/8        time:   [40.682 ns 40.706 ns 40.731 ns]
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
from_str_first/90       time:   [57.174 ns 57.217 ns 57.263 ns]
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
from_str_first/972      time:   [230.03 ns 230.16 ns 230.30 ns]
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  6 (6.00%) high severe

from_str_dupe/0         time:   [7.1614 ns 7.1690 ns 7.1782 ns]
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) high mild
  18 (18.00%) high severe
from_str_dupe/1         time:   [13.662 ns 13.676 ns 13.688 ns]
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) low mild
  1 (1.00%) high mild
from_str_dupe/7         time:   [6.9212 ns 7.0017 ns 7.0940 ns]
Found 22 outliers among 100 measurements (22.00%)
  1 (1.00%) high mild
  21 (21.00%) high severe
from_str_dupe/8         time:   [15.685 ns 15.695 ns 15.707 ns]
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe
from_str_dupe/90        time:   [25.798 ns 25.810 ns 25.823 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe
from_str_dupe/972       time:   [117.82 ns 117.90 ns 117.99 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

clone/0                 time:   [3.7363 ns 3.7436 ns 3.7532 ns]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe
clone/1                 time:   [3.7287 ns 3.7315 ns 3.7347 ns]
Found 7 outliers among 100 measurements (7.00%)
  7 (7.00%) high mild
clone/7                 time:   [3.7283 ns 3.7307 ns 3.7335 ns]
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe
clone/8                 time:   [4.5004 ns 4.5253 ns 4.5533 ns]
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe
clone/90                time:   [4.5612 ns 4.5752 ns 4.5895 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
clone/972               time:   [4.5107 ns 4.5211 ns 4.5331 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild

drop_dupe/0             time:   [30.779 ns 30.804 ns 30.834 ns]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  5 (5.00%) high severe
drop_dupe/1             time:   [30.766 ns 30.793 ns 30.826 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
drop_dupe/7             time:   [30.797 ns 30.822 ns 30.847 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
drop_dupe/8             time:   [34.222 ns 34.245 ns 34.269 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
drop_dupe/90            time:   [34.213 ns 34.242 ns 34.275 ns]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
drop_dupe/972           time:   [34.174 ns 34.192 ns 34.215 ns]
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  3 (3.00%) high severe

drop_last/0             time:   [30.707 ns 30.736 ns 30.776 ns]
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe
drop_last/1             time:   [30.730 ns 30.749 ns 30.771 ns]
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe
drop_last/7             time:   [30.760 ns 30.783 ns 30.810 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
drop_last/8             time:   [51.068 ns 51.135 ns 51.206 ns]
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  1 (1.00%) high severe
drop_last/90            time:   [58.003 ns 58.073 ns 58.147 ns]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
drop_last/972           time:   [141.16 ns 141.25 ns 141.33 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

as_str/0                time:   [311.14 ps 311.51 ps 311.91 ps]
Found 18 outliers among 100 measurements (18.00%)
  11 (11.00%) high mild
  7 (7.00%) high severe
as_str/1                time:   [311.05 ps 311.70 ps 312.46 ps]
Found 15 outliers among 100 measurements (15.00%)
  3 (3.00%) high mild
  12 (12.00%) high severe
as_str/7                time:   [310.55 ps 310.92 ps 311.31 ps]
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) high mild
  13 (13.00%) high severe
as_str/8                time:   [311.22 ps 312.05 ps 313.03 ps]
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) low mild
  16 (16.00%) high severe
as_str/90               time:   [310.33 ps 310.66 ps 311.08 ps]
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) high mild
  10 (10.00%) high severe
as_str/972              time:   [311.14 ps 311.84 ps 312.64 ps]
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  13 (13.00%) high severe

ahash/0                 time:   [31.021 ns 31.038 ns 31.057 ns]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
ahash/1                 time:   [31.061 ns 31.088 ns 31.118 ns]
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
ahash/7                 time:   [31.010 ns 31.032 ns 31.056 ns]
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  5 (5.00%) high mild
ahash/8                 time:   [31.024 ns 31.049 ns 31.075 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high severe
ahash/90                time:   [31.010 ns 31.038 ns 31.069 ns]
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
ahash/972               time:   [31.085 ns 31.108 ns 31.133 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high mild

eq/0                    time:   [310.87 ps 311.27 ps 311.77 ps]
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe
eq/1                    time:   [310.28 ps 310.59 ps 310.96 ps]
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  9 (9.00%) high severe
eq/7                    time:   [310.77 ps 311.12 ps 311.53 ps]
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  12 (12.00%) high severe
eq/8                    time:   [310.99 ps 311.31 ps 311.74 ps]
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  5 (5.00%) high mild
  5 (5.00%) high severe
eq/90                   time:   [311.74 ps 312.06 ps 312.40 ps]
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low mild
  9 (9.00%) high mild
  3 (3.00%) high severe
eq/972                  time:   [311.10 ps 311.49 ps 311.93 ps]
Found 10 outliers among 100 measurements (10.00%)
  8 (8.00%) high mild
  2 (2.00%) high severe
```
</details>